fix kotlin import (#7028)
