+++ core-kotlin-2/src/main/kotlin/com/baeldung/jvmannotations/Document.kt
@@ -9,19 +9,3 @@ interface Document {
 
     fun getType() = "document"
 }
-
-class TextDocument : Document {
-    override fun getType() = "text"
-
-    fun transformList(list : List<Number>) : List<Number> {
-        return list.filter { n -> n.toInt() > 1 }
-    }
-
-    fun transformListInverseWildcards(list : List<@JvmSuppressWildcards Number>) : List<@JvmWildcard Number> {
-        return list.filter { n -> n.toInt() > 1 }
-    }
-
-    var list : List<@JvmWildcard Any> = ArrayList()
-}
-
-class XmlDocument(d : Document) : Document by d
