+++ core-kotlin-2/src/main/kotlin/com/baeldung/jvmannotations/XmlDocument.kt
@@ -1,0 +1,5 @@
+package com.baeldung.jvmannotations
+
+import java.util.*
+
+class XmlDocument(d : Document) : Document by d
